Build liblldb 3.8 from sources (#41)
The OpenSUSE 42.3 doesn't have official package liblldb-devel that
we need to build lldb plugin for the lldb that's available for this
distro. So we build the devel library from sources.
This is only build time library, the lldb package contains the runtime
library, so the plugin can work without any extra work on the user
side.