Add locales to ubuntu 16.0.4
The MSBuild `Exec` task will use `export LC_ALL=en_US.UTF-8` as a
part of the `bash` invocation. When ubuntu lacks the `locales`
package this will cause a warning to be emitted from the `Exec`
usage and the outer MSBuild invocation. Adding the `locales`
package here to get our builds warning free as this makes it
significantly easier to manage our build reliability

https://github.com/microsoft/msbuild/issues/4194
https://github.com/dotnet/runtime/issues/34280