Change all of the tags that are "isLocal" to use the -local suffix
We wanted to be able to make images published by production to use a suffix-less tag, but that caused a conflict with the local tags for some dockerfiles, since they are already suffix-less. This change add -local to those local tags, and updates all of the dockerfiles that were using the local images.